ExamApp/requirements.txt
Django
django-cors-headers
djangorestframework
Pillow
psycopg2-binary
python-dotenv